+++ scripts/deployments/tools/deterministic-get-salt-zeta-connector.ts
