+++ scripts/deployments/tools/deterministic-get-salt-zeta-token.ts
