+++ scripts/tools/update-zeta-connector.ts
@@ -1,8 +1,8 @@
 import { isNetworkName } from "@zetachain/addresses";
 import { ethers, network } from "hardhat";
 
-import { getAddress } from "../../../lib/address.helpers";
-import { getZetaFactoryNonEth, isEthNetworkName } from "../../../lib/contracts.helpers";
+import { getAddress } from "../../lib/address.helpers";
+import { getZetaFactoryNonEth, isEthNetworkName } from "../../lib/contracts.helpers";
 
 async function updateZetaConnector() {
   if (!isNetworkName(network.name)) {
